backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pandas
openpyxl
python-multipart
python-dotenv
pyodbc
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
